# gamemaster-ai/docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: GameMaster AI API
  version: 0.2.0
  description: Auto-synced API surface generated from app/api route handlers.
servers:
  - url: /api
paths:
...
  /characters/{id}/inventory:
    get:
      summary: GET /api/characters/[id]/inventory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
    post:
      summary: POST /api/characters/[id]/inventory
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
  /characters/{id}/inventory/{itemId}:
    get:
      summary: GET /api/characters/[id]/inventory/[itemId]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: itemId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
    put:
      summary: PUT /api/characters/[id]/inventory/[itemId]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: itemId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
    delete:
      summary: DELETE /api/characters/[id]/inventory/[itemId]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: itemId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '429':
          description: Too Many Requests
        '500':
          description: Server Error
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token